social_media_api/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
click==8.1.8
dj-database-url==2.3.0
Django==5.1.7
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.0
uvicorn==0.34.0
whitenoise==6.9.0
